(Transactional)ActionEventEmitterEventStore wraps the original implementation
see: https://github.com/prooph/event-store/issues/229
