--- src/Exception/TransactionNotCommitted.php
@@ -1,18 +1,0 @@
-<?php
-/**
- * This file is part of the prooph/event-store.
- * (c) 2014-2016 prooph software GmbH <contact@prooph.de>
- * (c) 2015-2016 Sascha-Oliver Prolic <saschaprolic@googlemail.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Prooph\EventStore\Exception;
-
-class TransactionNotCommitted extends RuntimeException
-{
-    protected $message = 'The transaction has not been committed.';
-}
