+++ src/InMemoryEventStore.php
@@ -13,8 +13,12 @@ declare(strict_types=1);
 namespace Prooph\EventStore;
 
 use ArrayIterator;
-use Prooph\Common\Event\ActionEvent;
-use Prooph\Common\Event\ActionEventEmitter;
+use Iterator;
+use Prooph\Common\Messaging\Message;
+use Prooph\EventStore\Exception\StreamExistsAlready;
+use Prooph\EventStore\Exception\StreamNotFound;
+use Prooph\EventStore\Exception\TransactionAlreadyStarted;
+use Prooph\EventStore\Exception\TransactionNotStarted;
 use Prooph\EventStore\Metadata\MetadataMatcher;
 use Prooph\EventStore\Metadata\Operator;
 use Prooph\EventStore\Projection\InMemoryEventStoreProjection;
@@ -25,8 +29,9 @@ use Prooph\EventStore\Projection\ProjectionOptions;
 use Prooph\EventStore\Projection\Query;
 use Prooph\EventStore\Projection\ReadModel;
 use Prooph\EventStore\Projection\ReadModelProjection;
+use Prooph\EventStore\Util\Assertion;
 
-final class InMemoryEventStore extends AbstractTransactionalActionEventEmitterEventStore
+final class InMemoryEventStore implements TransactionalEventStore
 {
     /**
      * @var array
@@ -41,217 +46,241 @@ final class InMemoryEventStore extends AbstractTransactionalActionEventEmitterEv
     /**
      * @var bool
      */
-    private $inTransaction = false;
+    private $isInTransaction = false;
 
-    /**
-     * @var ActionEventEmitter
-     */
-    protected $actionEventEmitter;
-
-    public function __construct(ActionEventEmitter $actionEventEmitter)
+    public function create(Stream $stream): void
     {
-        $this->actionEventEmitter = $actionEventEmitter;
+        $streamName = $stream->streamName();
+        $streamNameString = $streamName->toString();
 
-        $actionEventEmitter->attachListener(self::EVENT_CREATE, function (ActionEvent $event): void {
-            $stream = $event->getParam('stream');
+        if (isset($this->streams[$streamNameString])
+            || isset($this->cachedStreams[$streamNameString])
+        ) {
+            throw StreamExistsAlready::with($streamName);
+        }
 
-            $streamName = $stream->streamName()->toString();
+        if ($this->isInTransaction) {
+            $this->cachedStreams[$streamNameString]['events'] = $stream->streamEvents();
+            $this->cachedStreams[$streamNameString]['metadata'] = $stream->metadata();
+        } else {
+            $this->streams[$streamNameString]['events'] = $stream->streamEvents();
+            $this->streams[$streamNameString]['metadata'] = $stream->metadata();
+        }
+    }
 
-            if (isset($this->streams[$streamName])
-                || isset($this->cachedStreams[$streamName])
+    public function appendTo(StreamName $streamName, Iterator $streamEvents): void
+    {
+        $streamNameString = $streamName->toString();
+
+        if (! isset($this->streams[$streamNameString])
+            && ! isset($this->cachedStreams[$streamNameString])
+        ) {
+            throw StreamNotFound::with($streamName);
+        }
+
+        if ($this->isInTransaction) {
+            if (! isset($this->cachedStreams[$streamNameString])) {
+                $this->cachedStreams[$streamNameString]['events'] = [];
+            }
+
+            foreach ($streamEvents as $streamEvent) {
+                $this->cachedStreams[$streamNameString]['events'][] = $streamEvent;
+            }
+        } else {
+            foreach ($streamEvents as $streamEvent) {
+                $this->streams[$streamNameString]['events'][] = $streamEvent;
+            }
+        }
+    }
+
+    public function load(
+        StreamName $streamName,
+        int $fromNumber = 1,
+        int $count = null,
+        MetadataMatcher $metadataMatcher = null
+    ): Stream {
+        Assertion::greaterOrEqualThan($fromNumber, 1);
+        Assertion::nullOrGreaterOrEqualThan($count, 1);
+
+        if (! isset($this->streams[$streamName->toString()])) {
+            throw StreamNotFound::with($streamName);
+        }
+
+        if (null === $metadataMatcher) {
+            $metadataMatcher = new MetadataMatcher();
+        }
+
+        $streamEvents = [];
+
+        foreach ($this->streams[$streamName->toString()]['events'] as $key => $streamEvent) {
+            /* @var Message $streamEvent */
+            if ($this->matchesMetadata($metadataMatcher, $streamEvent->metadata())
+                && ((null === $count
+                        && ($key + 1) >= $fromNumber
+                    ) || (null !== $count
+                        && ($key + 1) >= $fromNumber
+                        && ($key + 1) <= ($fromNumber + $count - 1)
+                    )
+                )
             ) {
-                return;
+                $streamEvents[] = $streamEvent;
             }
+        }
 
-            if ($this->inTransaction) {
-                $this->cachedStreams[$streamName]['events'] = $stream->streamEvents();
-                $this->cachedStreams[$streamName]['metadata'] = $stream->metadata();
-            } else {
-                $this->streams[$streamName]['events'] = $stream->streamEvents();
-                $this->streams[$streamName]['metadata'] = $stream->metadata();
-            }
+        if (empty($streamEvents)) {
+            throw StreamNotFound::with($streamName);
+        }
 
-            $event->setParam('result', true);
-        });
+        return new Stream(
+            $streamName,
+            new ArrayIterator($streamEvents),
+            $this->streams[$streamName->toString()]['metadata']
+        );
+    }
 
-        $actionEventEmitter->attachListener(self::EVENT_APPEND_TO, function (ActionEvent $event): void {
-            $streamName = $event->getParam('streamName')->toString();
-            $streamEvents = $event->getParam('streamEvents');
+    public function loadReverse(
+        StreamName $streamName,
+        int $fromNumber = PHP_INT_MAX,
+        int $count = null,
+        MetadataMatcher $metadataMatcher = null
+    ): Stream {
+        Assertion::greaterOrEqualThan($fromNumber, 1);
+        Assertion::nullOrGreaterOrEqualThan($count, 1);
 
-            if (! isset($this->streams[$streamName])
-                && ! isset($this->cachedStreams[$streamName])
+        if (! isset($this->streams[$streamName->toString()])) {
+            throw StreamNotFound::with($streamName);
+        }
+
+        if (null === $metadataMatcher) {
+            $metadataMatcher = new MetadataMatcher();
+        }
+
+        $streamEvents = [];
+
+        foreach ($this->streams[$streamName->toString()]['events'] as $key => $streamEvent) {
+            /* @var Message $streamEvent */
+            if ($this->matchesMetadata($metadataMatcher, $streamEvent->metadata())
+                && ((null === $count
+                        && ($key + 1) <= $fromNumber
+                    ) || (null !== $count
+                        && ($key + 1) <= $fromNumber
+                        && ($key + 1) >= ($fromNumber - $count + 1)
+                    )
+                )
             ) {
-                return;
+                $streamEvents[] = $streamEvent;
             }
+        }
 
-            if ($this->inTransaction) {
-                if (! isset($this->cachedStreams[$streamName])) {
-                    $this->cachedStreams[$streamName]['events'] = [];
-                }
+        if (empty($streamEvents)) {
+            throw StreamNotFound::with($streamName);
+        }
 
-                foreach ($streamEvents as $streamEvent) {
-                    $this->cachedStreams[$streamName]['events'][] = $streamEvent;
-                }
+        $streamEvents = array_reverse($streamEvents);
+
+        return new Stream(
+            $streamName,
+            new ArrayIterator($streamEvents),
+            $this->streams[$streamName->toString()]['metadata']
+        );
+    }
+
+    public function delete(StreamName $streamName): void
+    {
+        $streamNameString = $streamName->toString();
+
+        if (isset($this->streams[$streamNameString])) {
+            unset($this->streams[$streamNameString]);
+        } else {
+            throw StreamNotFound::with($streamName);
+        }
+    }
+
+    public function hasStream(StreamName $streamName): bool
+    {
+        return isset($this->streams[$streamName->toString()]);
+    }
+
+    public function fetchStreamMetadata(StreamName $streamName): array
+    {
+        if (! isset($this->streams[$streamName->toString()])) {
+            throw StreamNotFound::with($streamName);
+        }
+
+        return $this->streams[$streamName->toString()]['metadata'];
+    }
+
+    public function updateStreamMetadata(StreamName $streamName, array $newMetadata): void
+    {
+        if (! isset($this->streams[$streamName->toString()])) {
+            throw StreamNotFound::with($streamName);
+        }
+
+        $this->streams[$streamName->toString()]['metadata'] = $newMetadata;
+    }
+
+    public function beginTransaction(): void
+    {
+        if ($this->isInTransaction) {
+            throw new TransactionAlreadyStarted();
+        }
+
+        $this->isInTransaction = true;
+    }
+
+    public function commit(): void
+    {
+        if (! $this->isInTransaction) {
+            throw new TransactionNotStarted();
+        }
+
+        foreach ($this->cachedStreams as $streamName => $data) {
+            if (isset($data['metadata'])) {
+                $this->streams[$streamName] = $data;
             } else {
-                foreach ($streamEvents as $streamEvent) {
+                foreach ($data['events'] as $streamEvent) {
                     $this->streams[$streamName]['events'][] = $streamEvent;
                 }
             }
+        }
 
-            $event->setParam('result', true);
-        });
+        $this->cachedStreams = [];
+        $this->isInTransaction = false;
+    }
 
-        $actionEventEmitter->attachListener(self::EVENT_LOAD, function (ActionEvent $event): void {
-            $streamName = $event->getParam('streamName');
-            $fromNumber = $event->getParam('fromNumber');
-            $count = $event->getParam('count');
-            $metadataMatcher = $event->getParam('metadataMatcher');
+    public function rollback(): void
+    {
+        if (! $this->isInTransaction) {
+            throw new TransactionNotStarted();
+        }
 
-            if (! isset($this->streams[$streamName->toString()])) {
-                return;
-            }
+        $this->cachedStreams = [];
+        $this->isInTransaction = false;
+    }
 
-            if (null === $metadataMatcher) {
-                $metadataMatcher = new MetadataMatcher();
-            }
+    public function isInTransaction(): bool
+    {
+        return $this->isInTransaction;
+    }
 
-            $streamEvents = [];
+    /**
+     * @throws \Exception
+     *
+     * @return mixed
+     */
+    public function transactional(callable $callable)
+    {
+        $this->beginTransaction();
 
-            foreach ($this->streams[$streamName->toString()]['events'] as $key => $streamEvent) {
-                if ($this->matchesMetadata($metadataMatcher, $streamEvent->metadata())
-                    && ((null === $count
-                            && ($key + 1) >= $fromNumber
-                        ) || (null !== $count
-                            && ($key + 1) >= $fromNumber
-                            && ($key + 1) <= ($fromNumber + $count - 1)
-                        )
-                    )
-                ) {
-                    $streamEvents[] = $streamEvent;
-                }
-            }
+        try {
+            $result = $callable($this);
+            $this->commit();
+        } catch (\Exception $e) {
+            $this->rollback();
+            throw $e;
+        }
 
-            $event->setParam('stream', new Stream(
-                $streamName,
-                new ArrayIterator($streamEvents),
-                $this->streams[$streamName->toString()]['metadata']
-            ));
-        });
-
-        $actionEventEmitter->attachListener(self::EVENT_LOAD_REVERSE, function (ActionEvent $event): void {
-            $streamName = $event->getParam('streamName');
-            $fromNumber = $event->getParam('fromNumber');
-            $count = $event->getParam('count');
-            $metadataMatcher = $event->getParam('metadataMatcher');
-
-            if (! isset($this->streams[$streamName->toString()])) {
-                return;
-            }
-
-            if (null === $metadataMatcher) {
-                $metadataMatcher = new MetadataMatcher();
-            }
-
-            $streamEvents = [];
-
-            foreach ($this->streams[$streamName->toString()]['events'] as $key => $streamEvent) {
-                if ($this->matchesMetadata($metadataMatcher, $streamEvent->metadata())
-                    && ((null === $count
-                            && ($key + 1) <= $fromNumber
-                        ) || (null !== $count
-                            && ($key + 1) <= $fromNumber
-                            && ($key + 1) >= ($fromNumber - $count + 1)
-                        )
-                    )
-                ) {
-                    $streamEvents[] = $streamEvent;
-                }
-            }
-
-            $streamEvents = array_reverse($streamEvents);
-
-            $event->setParam('stream', new Stream(
-                $streamName,
-                new ArrayIterator($streamEvents),
-                $this->streams[$streamName->toString()]['metadata']
-            ));
-        });
-
-        $actionEventEmitter->attachListener(self::EVENT_DELETE, function (ActionEvent $event): void {
-            $streamName = $event->getParam('streamName')->toString();
-
-            if (isset($this->streams[$streamName])) {
-                unset($this->streams[$streamName]);
-                $event->setParam('result', true);
-            } else {
-                $event->setParam('result', false);
-            }
-        });
-
-        $actionEventEmitter->attachListener(self::EVENT_HAS_STREAM, function (ActionEvent $event): void {
-            $streamName = $event->getParam('streamName');
-
-            $result = isset($this->streams[$streamName->toString()]);
-
-            $event->setParam('result', $result);
-        });
-
-        $actionEventEmitter->attachListener(self::EVENT_FETCH_STREAM_METADATA, function (ActionEvent $event): void {
-            $streamName = $event->getParam('streamName');
-
-            if (! isset($this->streams[$streamName->toString()])) {
-                return;
-            }
-
-            $metadata = $this->streams[$streamName->toString()]['metadata'];
-
-            $event->setParam('metadata', $metadata);
-        });
-
-        $actionEventEmitter->attachListener(self::EVENT_UPDATE_STREAM_METADATA, function (ActionEvent $event): void {
-            $streamName = $event->getParam('streamName');
-            $metadata = $event->getParam('metadata');
-
-            if (! isset($this->streams[$streamName->toString()])) {
-                $event->setParam('result', false);
-
-                return;
-            }
-
-            $this->streams[$streamName->toString()]['metadata'] = $metadata;
-
-            $event->setParam('result', true);
-        });
-
-        $actionEventEmitter->attachListener(self::EVENT_BEGIN_TRANSACTION, function (ActionEvent $event): void {
-            $this->inTransaction = true;
-
-            $event->setParam('inTransaction', true);
-        });
-
-        $actionEventEmitter->attachListener(self::EVENT_COMMIT, function (ActionEvent $event): void {
-            foreach ($this->cachedStreams as $streamName => $data) {
-                if (isset($data['metadata'])) {
-                    $this->streams[$streamName] = $data;
-                } else {
-                    foreach ($data['events'] as $streamEvent) {
-                        $this->streams[$streamName]['events'][] = $streamEvent;
-                    }
-                }
-            }
-
-            $this->cachedStreams = [];
-            $this->inTransaction = false;
-
-            $event->setParam('inTransaction', false);
-        });
-
-        $actionEventEmitter->attachListener(self::EVENT_ROLLBACK, function (ActionEvent $event): void {
-            $this->cachedStreams = [];
-            $this->inTransaction = false;
-
-            $event->setParam('inTransaction', false);
-        });
+        return $result ?: true;
     }
 
     public function createQuery(): Query
